extract fraction to own file
